Move all components from app to relevant modules
