+++ client/src/app/app.component.html
@@ -1,3 +1,1 @@
-<!--<app-search></app-search>-->
-<!--<app-books-list></app-books-list>-->
 <router-outlet></router-outlet>
