+++ client/src/app/app.module.ts
@@ -1,22 +1,13 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { HttpClientModule } from '@angular/common/http';
 import { RouterModule } from '@angular/router';
-
-import { MatInputModule } from '@angular/material/input';
-import { MatIconModule } from '@angular/material/icon';
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
-
-import { FlexLayoutModule } from '@angular/flex-layout';
+import { CommonModule } from '@angular/common';
 
 import { StoreModule } from '@ngrx/store';
-import { EffectsModule } from '@ngrx/effects';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
-
-import { ImageDirective } from './directives/image.directive';
+import { EffectsModule } from '@ngrx/effects';
 
 import { AppComponent } from './app.component';
 
@@ -24,35 +15,18 @@ import { AppRoutingModule } from './app-routing.module';
 
 import { environment } from '../environments/environment';
 import { reducers } from './store/state.index';
-import { BooksEffect } from './store/books.effect';
-
-import { SearchComponent } from './components/search/search.component';
-import { BooksListComponent } from './components/books-list/books-list.component';
-import { BookCardComponent } from './components/book-card/book-card.component';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    SearchComponent,
-    BooksListComponent,
-    ImageDirective,
-    BookCardComponent
-  ],
+  declarations: [ AppComponent ],
   imports: [
     AppRoutingModule,
-    HttpClientModule,
     BrowserModule,
     BrowserAnimationsModule,
-    FlexLayoutModule,
-    MatInputModule,
+    CommonModule,
+    HttpClientModule,
     StoreModule.forRoot(reducers),
-    EffectsModule.forRoot([BooksEffect]),
+    EffectsModule.forRoot(),
     StoreDevtoolsModule.instrument({ maxAge: 25, logOnly: environment.production }),
-    FormsModule,
-    MatIconModule,
-    MatButtonModule,
-    ReactiveFormsModule,
-    MatCardModule,
     RouterModule
   ],
   providers: [],
