+++ client/src/app/feed/components/book-card/book-card.component.ts
