+++ client/src/app/feed/components/feed/feed.component.html
@@ -1,1 +1,2 @@
-<p>feed works!</p>
+<app-search></app-search>
+<app-books-list></app-books-list>
