+++ client/src/app/feed/directives/image.directive.ts
