+++ client/src/app/feed/models/books.model.ts
