+++ client/src/app/search/components/search/search.component.scss
